% Auto-generated: do not edit by hand
\name{''DongjakDashComponents2}

\alias{''DongjakDashComponents2}

\title{DongjakDashComponents2 component}

\description{
ExampleComponent is an example component. It takes a property, `label`, and displays it. It renders an input with the property `value` which is editable by the user.
}

\usage{
''DongjakDashComponents2(id=NULL, label=NULL, test=NULL, value=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{label}{Character. A label that will be printed when this component is rendered.}

\item{test}{Dash component. }

\item{value}{Character. The value displayed in the input.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

