% Auto-generated: do not edit by hand
\name{''DongjakDashComponents2}

\alias{''DongjakDashComponents2}

\title{DongjakDashComponents2 component}

\description{
Component description
}

\usage{
''DongjakDashComponents2(id=NULL, test=NULL)
}

\arguments{
\item{id}{Character. Unique ID to identify this component in Dash callbacks.}

\item{test}{A list of or a singular dash component, string or number. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

